feature: Implemented Pathway Power. You can now use Blueprints and Blueprint Palettes with the Pathway Power to generate a path from point A to point B. The Blueprint pieces must have Entrances marked on the edges with the Entrance Power. Pathways can only be made when the pieces provided have adequate Entrances marked.
To connect to each other, an Entrance must have the same dimensions as the other piece's Entrance.

If X is a block, and E is an entrance, a corner piece such as:

XXX
EXX
XEX

Would allow you to create a 2D pathway. However, a piece with straight horizontal entrances such as:

XEX
XXX
XEX

Would not (by itself) create a Pathway, since it needs to be able to go in the opposite direction, but it cannot connect to itself when rotated. A corner piece can, however.

To create 3D pathways that go up and down, you must have vertical pieces with an entrance on top and bottom.

Ideally, combining all of these pieces into one Blueprint Palette will create a very organic-looking Pathway when used with the Pathway Power, since it has many pieces at its disposal to pathfind from point A to B.